refactor: replace modals in folders
