+++ src/components/AdoptionModal/AdoptionModal.module.scss
@@ -22,7 +22,7 @@
 	@include small-desktop {
 		width: 360px;
 		height: 532px;
-		background-image: url('../../assets//modal/adoption/1280_1x.webp');
+		background-image: url('../../assets/modal/adoption/1280_1x.webp');
 	}
 	@include large-desktop {
 		width: 360px;
