+++ src/components/ThanksModal/ThanksModal.tsx
