+++ src/layout/HomeLayout/HomeLayout.tsx
@@ -1,12 +1,13 @@
-import Modal from '../../components/Modal/Modal';
-// import ThanksModal from '../../components/Modal/ThanksModal';
+import Modal from '../ModalLayout/Modal';
+import ThanksModal from '../../components/ThanksModal/ThanksModal';
 import Footer from '../Footer/Footer';
 import Header from '../Header/Header';
 import { Outlet } from 'react-router-dom';
 // import { useContext } from 'react';
 // import { ModalContext } from '../../context/ModalContext';
 import { useModalContext } from '../../context/useGlobalContext';
-import AdoptionModal from '../../components/Modal/AdoptionModal';
+import AdoptionModal from '../../components/AdoptionModal/AdoptionModal';
+import ContactModal from '../../components/ContactModal/ContactModal';
 
 const HomeLayout = () => {
 	// const modalContext = useModalContext();
@@ -20,14 +21,30 @@ const HomeLayout = () => {
 			<Header />
 			<Outlet />
 			<Footer />
-			<Modal
+		{/*	<ContactModal
 				isModal={isModalOpen}
 				openModal={openModal}
 				closeModal={closeModal}
 			>
-				{/* <ThanksModal /> */}
-				<AdoptionModal />
-			</Modal>
+
+				<ContactModal />
+			</ContactModal>*/}
+
+		{/*	<ContactModal
+				isModal={isModalOpen}
+				openModal={openModal}
+				closeModal={closeModal}
+			>
+				 <ThanksModal />
+			</ContactModal>*/}
+
+			{/*<Modal
+				isModal={isModalOpen}
+				openModal={openModal}
+				closeModal={closeModal}
+			>
+				<ContactModal />
+			</Modal>*/}
 		</>
 	);
 };
